+++ src/test/java/seedu/address/logic/commands/event/EditEventCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.event;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -18,7 +18,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.event.EditEventCommand;
+import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.event.EditEventCommand.EditEventDescriptor;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
